refactor: Validator 클래스 생성 및 검증 메서드 이동
- 로또 번호 및 보너스 번호, 구입 금액 검증 메서드를 Validator 클래스로 이동
- 다른 클래스에서 중복되던 검증 로직 제거하여 코드 모듈화 및 재사용성 향상
- 코드 정리 및 유지보수성 개선
