+++ src/main/java/lotto/LottoController.java
@@ -5,16 +5,10 @@ import java.util.stream.Stream;
 
 public class LottoController {
 
-    public static void validatePurchaseNumber(int amount) {
-        if (amount <= 0 || amount % 1000 != 0) {
-            throw new IllegalArgumentException("[ERROR] 구입금액은 1000원단위로 입력가능합니다.");
-        }
-    }
-
     public int parsePurchaseNumber(String input) {
         try {
             int purchaseNumber = Integer.parseInt(input.trim());
-            validatePurchaseNumber(purchaseNumber);
+            Validator.validatePurchaseNumber(purchaseNumber);
             return purchaseNumber / 1000;
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("[ERROR] 잘못된 형식의 구입금액 입니다.");
@@ -27,7 +21,7 @@ public class LottoController {
                     .map(String::trim)
                     .map(Integer::parseInt)
                     .toList();
-            Lotto.validate(numbers);
+            Validator.validateWinNumber(numbers);
             return numbers;
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("[ERROR] 잘못된 형식의 로또 번호입니다.");
